requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0

sqlalchemy==2.0.38
psycopg[binary]==3.2.4
alembic==1.14.1

pydantic-settings==2.7.1
pyyaml==6.0.2

httpx==0.28.1
